added handle plus and minus icon with functionality

+++ src/Components/Cart/ProductInfo.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import {deleteFromCart} from '../../Redux/Actions/cartItemAction'
+import {deleteFromCart, addToCart} from '../../Redux/Actions/cartItemAction'
 import {Form, Input, Segment, Image, Button, Grid, Item, Icon, Header} from 'semantic-ui-react'
 import {connect} from 'react-redux'
 
@@ -7,8 +7,40 @@ import {connect} from 'react-redux'
 class ProductInfo extends Component {
 
   handleDelete=()=>{
-    // console.log('click me');
-    fetch(`http://localhost:4000/cartitem/delete/${this.props.cartItem.itemInfo.id}`, {
+    fetch(`http://localhost:4000/cartitem/remove/${this.props.cartItem.item_id}`, {
+      method : 'DELETE',
+      headers: {
+          'Authorization': `bearer ${localStorage.token}`
+      }
+    })
+    .then(r => r.json())
+    .then((data) => {
+      this.props.deleteFromCart(data.cart_item.id)
+    })
+  }
+
+  handlePlus=()=>{
+    let item_id = this.props.cartItem.item_id
+    fetch(`http://localhost:4000/cart_items`, {
+      method : "POST",
+      headers: {
+        'content-type' : 'application/json'
+      },
+      body: JSON.stringify( {
+        quantity: 1,
+        item_id,
+        cart_id: this.props.cartItem.cart_id
+      })
+    }
+  )
+  .then(r => r.json())
+  .then((data) => {
+    this.props.addToCart(data);
+    })
+  }
+
+  handleMinus=()=>{
+    fetch(`http://localhost:4000/cartitem/delete/${this.props.cartItem.id}`, {
       method : 'DELETE',
       headers: {
           'Authorization': `bearer ${localStorage.token}`
@@ -17,16 +49,18 @@ class ProductInfo extends Component {
     .then(r => r.json())
     .then((data) => {
       console.log(data);
-       // debugger
-      //was about to dispatch the action
-      this.props.deleteFromCart(data.cart_item.id)
-
+      if (data.cart_item_deleted){
+        this.props.deleteFromCart(data.cart_item.id)
+      } else {
+        this.props.addToCart(data)
+      }
     })
   }
 
   render() {
   // debugger
     let {size, name, price, image, color} = this.props.cartItem.item
+    let {quantity} = this.props.cartItem
     // console.log(this.props.cartItem);
      // console.log(quantity)
     return (
@@ -36,11 +70,16 @@ class ProductInfo extends Component {
         <Item.Header>{name}</Item.Header>
           <Item.Meta>
             <span className='color'>color: <b>{color}</b></span>
+              <Item.Meta>
+                <span className='quantity'>quantity: <b>{quantity}</b></span>
+              </Item.Meta>
           </Item.Meta>
        <Item.Meta>
          <span className='price'>${price}</span>
        </Item.Meta>
-       <Button style={{margin: 0}} icon="plus"/><Input className="item-count"/><Button icon="minus"/><br/>
+         <Button onClick={this.handlePlus} style={{margin: 0}} icon="plus"/>
+          <Input className="item-count">{quantity}</Input>
+         <Button onClick={this.handleMinus} icon="minus"/><br/>
         <div>
          <Button icon="trash alternate outline" onClick={this.handleDelete} />
          <span size="small">Remove Item</span>
@@ -53,6 +92,6 @@ class ProductInfo extends Component {
 
 }
 
-export default connect(null, {deleteFromCart})(ProductInfo);
+export default connect(null, {deleteFromCart, addToCart})(ProductInfo);
 
 //need to send cartItem id to the backend
